requirements.txt
requests==2.31.0
dnspython==2.4.2
python-nmap==0.7.1
reportlab==3.6.10
paramiko==3.4.0
beautifulsoup4==4.12.3
scapy==2.5.0
tqdm==4.66.1
termcolor==2.4.0